pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "youtube-finance-ai"
version = "0.1.0"
description = "YouTube媒体下载和财经AI分析工具"
requires-python = ">=3.8"
dependencies = [
    "yt-dlp>=2024.8.6",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
]
whisper = [
    "openai-whisper>=20231117",
]
gemini = [
    "google-genai>=0.8.0",
    "google-api-core>=2.15.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
]
